chunks meshing fixed

+++ src/main/java/com/ohussar/VoxelEngine/Blocks/Block.java
@@ -3,9 +3,9 @@ package com.ohussar.VoxelEngine.Blocks;
 import org.lwjgl.util.vector.Vector3f;
 
 public class Block {
-    public int blockType;
+    public byte blockType;
     public Vector3f pos;
-    public Block(int blockType, Vector3f pos){
+    public Block(byte blockType, Vector3f pos){
         this.blockType = blockType;
         this.pos = pos;
     }
